# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HostMyImg API",
        "description": "API for hosting imagens (.png, .jpg, .gif) online",
        "termsOfService": "http://localhost:3000/terms",
        "version": "1.0.0",
        "contact": {
            "email": "samantafluture@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/"
        }
    ],
...
    "paths": {
        "/users/free": {
            "get": {
                "tags": ["users"],
                "summary": "List all users with no authentication",
                "description": "Get all users without needing the authentication token",
                "operationId": "getAllUsers",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Error Server"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": ["users"],
                "summary": "List all users",
                "description": "Get all users created through the singup route",
                "operationId": "getAllUsers",
                "produces": ["application/json"],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not Authorized!"
                    },
                    "500": {
                        "description": "Internal Error Server"
                    }
                }
            }
        },
        "/users/{_id}": {
            "get": {
                "tags": ["users"],
                "summary": "List user by ID",
                "description": "Get all data from a user while passing its ID by params",
                "operationId": "getUserById",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true
                    }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized!"
                    },
                    "500": {
                        "description": "Internal Error Server"
                    }
                }
            },
            "delete": {
                "tags": ["users"],
                "summary": "Delete user by ID",
                "description": "Remove all data from a user while passing its ID by params",
                "operationId": "deleteUser",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true
                    }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Not Authorized!"
                    },
                    "500": {
                        "description": "Internal Error Server"
                    }
                }
            }
        },
        "/users/{_id}/posts": {
            "get": {
                "tags": ["users"],
                "summary": "List posts by user",
                "description": "Get all posts from a user while passing the user ID by params",
                "operationId": "getPostsByUser",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true
                    }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized!"
                    },
                    "500": {
                        "description": "Internal Error Server"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "createAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "posts": {
                        "type": "array",
                        "items": {
                            "_id": {
                                "type": "string"
                            }
                        }
                    },
                    "_v": {
                        "type": "integer"
                    }
                }
            },
            "Post": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "size": {
                        "type": "number"
                    },
                    "url": {
                        "type": "string"
                    },
                    "createAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "type": "string"
                    },
                    "_v": {
                        "type": "integer"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-access-token"
            }
        }
    }
}